backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
awsebcli==3.20.5
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blessed==1.20.0
botocore==1.29.81
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==39.0.2
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
idna==3.4
jmespath==1.0.1
jsonschema==3.2.0
marshmallow==3.19.0
packaging==23.0
paramiko==3.0.0
pathspec==0.10.1
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
stripe==5.2.0
termcolor==1.1.0
texttable==1.6.7
tomli==2.0.1
urllib3==1.26.14
wcwidth==0.1.9
websocket-client==0.59.0
